@@ -1,37 +1,29 @@
 {
-  "name": "vue_tsx",
+  "name": "vite-project",
+  "private": true,
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vuedx-typecheck . && vite build"
+    "build": "vue-tsc && vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "ant-design-vue": "^2.1.6",
-    "vue": "3.0.11",
-    "vue-router": "^4.0.6",
-    "vuex": "4.0.0"
+    "ant-design-vue": "^4.1.2",
+    "vue": "3.4.21",
+    "vue-router": "^4.3.0",
+    "vuex": "4.1.0"
   },
   "devDependencies": {
-    "@types/node": "^14.14.41",
-    "@typescript-eslint/eslint-plugin": "^4.22.0",
-    "@typescript-eslint/parser": "^4.22.0",
-    "@vitejs/plugin-vue": "^1.2.1",
-    "@vitejs/plugin-vue-jsx": "^1.1.3",
-    "@vue/compiler-sfc": "3.0.11",
-    "@vuedx/typecheck": "^0.6.3",
-    "@vuedx/typescript-plugin-vue": "^0.6.3",
-    "eslint": "^7.24.0",
-    "eslint-config-prettier": "^8.2.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-vue": "7.9.0",
-    "less": "^4.1.1",
-    "prettier": "^2.2.1",
-    "sass": "^1.32.11",
-    "typescript": "^4.2.4",
-    "vite": "2.2.1",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vitejs/plugin-vue-jsx": "^3.1.0",
+    "sass": "^1.72.0",
+    "typescript": "^5.4.3",
+    "unplugin-vue-components": "^0.26.0",
+    "vite": "5.2.6",
     "vite-plugin-style-import": "^0.10.0",
-    "vue-eslint-parser": "^7.6.0",
     "vue-socket.io": "^3.0.10",
-    "webrtc-adapter": "^8.0.0"
+    "vue-tsc": "^2.0.7",
+    "webrtc-adapter": "^9.0.0"
   }
 }
