@@ -1,63 +1,20 @@
-import path from 'path';
-import { defineConfig, UserConfigExport } from 'vite';
-import vue from '@vitejs/plugin-vue';
-import vueJsx from '@vitejs/plugin-vue-jsx';
-import styleImport from 'vite-plugin-style-import';
+import { defineConfig } from 'vite'
+import vue from '@vitejs/plugin-vue'
+import vueJsx from '@vitejs/plugin-vue-jsx'
+import Components from 'unplugin-vue-components/vite'
+import { AntDesignVueResolver } from 'unplugin-vue-components/resolvers'
 
-const pathResolve = (pathStr: string) => {
-  return path.resolve(__dirname, pathStr);
-};
-
-// export default (): UserConfigExport => {
-//   return {
-//     css: {
-//       preprocessorOptions: {
-//         less: {
-//           javascriptEnabled: true,
-//         },
-//       },
-//     },
-//     plugins: [
-//       vue(),
-//       vueJsx(),
-//       styleImport({
-//         libs: [
-//           {
-//             libraryName: 'ant-design-vue',
-//             esModule: true,
-//             resolveStyle: (name) => {
-//               return `ant-design-vue/es/${name}/style/index`;
-//             },
-//           },
-//         ],
-//       }),
-//     ],
-//   };
-// };
-module.exports = defineConfig({
-  css: {
-    preprocessorOptions: {
-      less: {
-        javascriptEnabled: true,
-      },
-    },
-  },
+// https://vitejs.dev/config/
+export default defineConfig({
   plugins: [
     vue(),
     vueJsx(),
-    styleImport({
-      libs: [
-        {
-          libraryName: 'ant-design-vue',
-          esModule: true,
-          resolveStyle: (name) => {
-            return `ant-design-vue/es/${name}/style/css`;
-          },
-        },
+    Components({
+      resolvers: [
+        AntDesignVueResolver({
+          importStyle: false, // css in js
+        }),
       ],
     }),
   ],
-  alias: {
-    '@': pathResolve('./src'),
-  },
-});
+})
