@@ -1,4 +1,5 @@
 import { defineComponent, ref, onMounted, toRaw, nextTick } from 'vue';
+import type { RouteLocationNormalizedLoaded } from 'vue-router'
 import { useRoute } from 'vue-router';
 import { Layout, Menu } from 'ant-design-vue';
 import './App.scss';
@@ -14,49 +15,45 @@ interface MenuItem {
 
 const App = defineComponent({
   name: 'App',
-
   setup() {
-    const menuItem = ref<MenuItem[]>([]);
+    const menuItem = ref<MenuItem[]>([
+      {
+        index: '/getUserMedia',
+        title: '获取设备',
+      },
+      {
+        index: '/mediaDevices',
+        title: '音视频数据采集',
+      },
+      {
+        index: '/3',
+        title: '录制实战',
+        subs: [
+          { index: '/3/mediaRecoder', title: '录制音视频' },
+          { index: '/3/getDisplayMedia', title: '屏幕捕获' },
+        ],
+      },
+      {
+        index: '/socketIO',
+        title: '信令-socketIO',
+      },
+      {
+        index: '/RTCPeerConnection',
+        title: '端对端1V1传输',
+      },
+    ]);
     const selectedKeys = ref<string[]>([]);
     const openKeys = ref<string[]>([]);
-    const route = useRoute();
+    const route = useRoute() as RouteLocationNormalizedLoaded
 
     onMounted(() => {
-      const item: MenuItem[] = [
-        {
-          index: '/getUserMedia',
-          title: '获取设备',
-        },
-        {
-          index: '/mediaDevices',
-          title: '音视频数据采集',
-        },
-        {
-          index: '/3',
-          title: '录制实战',
-          subs: [
-            { index: '/3/mediaRecoder', title: '录制音视频' },
-            { index: '/3/getDisplayMedia', title: '屏幕捕获' },
-          ],
-        },
-        {
-          index: '/socketIO',
-          title: '信令-socketIO',
-        },
-        {
-          index: '/RTCPeerConnection',
-          title: '端对端1V1传输',
-        },
-      ];
-      menuItem.value = item;
-
       nextTick(() => {
-        const { path } = toRaw(route) as any;
+        const { path } = toRaw(route);
 
         setTimeout(() => {
-          const result = item.filter((i) => path.value.includes(i.index));
+          const result = menuItem.value.filter((i) => path.includes(i.index));
           if (result[0]?.subs?.length) {
-            const _subs = result[0].subs.filter((i) => path.value.includes(i.index));
+            const _subs = result[0].subs.filter((i) => path.includes(i.index));
             selectedKeys.value.push(_subs[0].index);
             openKeys.value.push(result[0].index);
           } else {
@@ -67,7 +64,7 @@ const App = defineComponent({
       });
     });
 
-    const ItemNode = (index: string | number | undefined, title: any) => {
+    const ItemNode = (index: string, title: string) => {
       return (
         <Item key={index}>
           <router-link to={index}>{title}</router-link>
